server/requirements/common.txt
Django==4.1.3
Pillow==9.3.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
python-dotenv==0.21.0
django-storages==1.13.1
boto3==1.26.17
gunicorn==20.1.0
requests==2.26.0
drf-yasg==1.21.4
django-redis==5.2.0
django-apscheduler==0.6.2
rembg==2.0.30
